+++ beacon-chain/sync/subscriber_beacon_blocks.go
@@ -6,12 +6,14 @@ import (
 	"os"
 	"path"
 
+	"github.com/ethereum/go-ethereum/crypto/kzg4844"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/blockchain"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/core/transition/interop"
 	"github.com/prysmaticlabs/prysm/v5/config/features"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/interfaces"
 	"github.com/prysmaticlabs/prysm/v5/io/file"
+	"github.com/prysmaticlabs/prysm/v5/time/slots"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -33,6 +35,8 @@ func (s *Service) beaconBlockSubscriber(ctx context.Context, msg proto.Message) 
 		return err
 	}
 
+	go s.attemptBlobSaveAndBroadcast(ctx, signed)
+
 	if err := s.cfg.chain.ReceiveBlock(ctx, signed, root, nil); err != nil {
 		if blockchain.IsInvalidBlock(err) {
 			r := blockchain.InvalidBlockRoot(err)
@@ -53,6 +57,61 @@ func (s *Service) beaconBlockSubscriber(ctx context.Context, msg proto.Message) 
 		return err
 	}
 	return err
+}
+
+// attemptBlobSaveAndBroadcast handles the process of saving and broadcasting blobs
+// by reconstructing blob sidecars from the execution layer (EL) and integrating them into the chain.
+func (s *Service) attemptBlobSaveAndBroadcast(ctx context.Context, block interfaces.ReadOnlySignedBeaconBlock) {
+	startTime, err := slots.ToTime(uint64(s.cfg.chain.GenesisTime().Unix()), block.Block().Slot())
+	if err != nil {
+		log.WithError(err).Error("Failed to convert slot to time")
+	}
+
+	blockRoot, err := block.Block().HashTreeRoot()
+	if err != nil {
+		log.WithError(err).Error("Failed to calculate block root")
+		return
+	}
+
+	// Reconstruct blob sidecars from the EL
+	blobSidecars, err := s.cfg.executionReconstructor.ReconstructBlobSidecars(ctx, block, blockRoot)
+	if err != nil {
+		log.WithError(err).Error("Failed to reconstruct blob sidecars")
+		return
+	}
+
+	// Get indices of already existing blobs in the database to avoid duplicates
+	indices, err := s.cfg.blobStorage.Indices(blockRoot)
+	if err != nil {
+		log.WithError(err).Error("Failed to retrieve indices for block")
+		return
+	}
+
+	// Iterate through the reconstructed sidecars, save them to the chain, and broadcast them to peers
+	for i, sidecar := range blobSidecars {
+		if indices[i] {
+			continue // Skip if the blob already exists in the database
+		}
+
+		if err := kzg4844.VerifyBlobProof(kzg4844.Blob(sidecar.Blob), kzg4844.Commitment(sidecar.KzgCommitment), kzg4844.Proof(sidecar.KzgProof)); err != nil {
+			log.WithFields(blobFields(sidecar.ROBlob)).WithError(err).Error("Failed to verify blob proof")
+			continue
+		}
+
+		if err := s.cfg.chain.ReceiveBlob(ctx, sidecar); err != nil {
+			log.WithFields(blobFields(sidecar.ROBlob)).WithError(err).Error("Failed to receive blob")
+		}
+
+		if err := s.cfg.p2p.BroadcastBlob(ctx, sidecar.Index, sidecar.BlobSidecar); err != nil {
+			log.WithFields(blobFields(sidecar.ROBlob)).WithError(err).Error("Failed to broadcast blob sidecar")
+		}
+
+		blobRecoveredFromELCount.Inc()
+
+		fields := blobFields(sidecar.ROBlob)
+		fields["sinceSlotStartTime"] = s.cfg.clock.Now().Sub(startTime)
+		log.WithFields(fields).Debug("Successfully received and processed blob sidecar from EL")
+	}
 }
 
 // WriteInvalidBlockToDisk as a block ssz. Writes to temp directory.